chore: Modify table name to contact
