+++ src/main/resources/db/migration/V1__TABLE_CREATION.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS users (
+CREATE TABLE IF NOT EXISTS contact (
 id INT PRIMARY KEY NOT NULL,
 first_name VARCHAR(255) NOT NULL,
 last_name VARCHAR(255) NOT NULL,
